refactor (resultView) : Refactor codes
