+++ src/main/java/lotto/controller/LottoGame.java
@@ -1,6 +1,6 @@
 package lotto.controller;
 
-import lotto.lottogenerator.LottoNumberPool;
+import lotto.model.LottoNumberPool;
 import lotto.model.Money;
 import lotto.model.MyLottos;
 import lotto.view.InputView;
