+++ src/main/java/lotto/model/LottoNumberPool.java
@@ -1,7 +1,6 @@
-package lotto.lottogenerator;
+package lotto.model;
 
-import lotto.model.LottoNumbers;
-import lotto.model.MyLottos;
+import lotto.lottogenerator.LottoNumGeneratorByAuto;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,13 +8,13 @@ import java.util.List;
 public class LottoNumberPool {
     private static List<LottoNumbers> myLottoNumbers = new ArrayList<>();
 
-    public static MyLottos getMyLottos(int autoLottoCount){
+    public static MyLottos getMyLottos(int autoLottoCount) {
         generateAutomatically(autoLottoCount);
         return new MyLottos(myLottoNumbers);
     }
 
     private static void generateAutomatically(int autoLottoCount) {
-        for(int i = 0; i < autoLottoCount ; i++){
+        for (int i = 0; i < autoLottoCount; i++) {
             myLottoNumbers.add(new LottoNumbers(LottoNumGeneratorByAuto.generate()));
         }
     }
